Rename polarity_plotting -> other_tools, move mag_angles to other_tools
